发红包

+++ src/main/groovy/com/ttpod/star/admin/PushController.groovy
@@ -47,7 +47,7 @@ class PushController extends BaseController {
             _id: { it == null ? "${Web.currentUserId}_${System.currentTimeMillis()}" as String: it as String},
             user_ids: {String ids-> if (StringUtils.isNotBlank(ids)) { ids.split(",").collect { it as Integer}}},
             text: Str, link_url:Str, img_url: Str, is_notify: {Boolean.valueOf((String)it)}, umeng_title: Str, umeng_text: Str,
-            umeng_event: Str, umeng_event_room_id: {StringUtils.isBlank(it as String) ? null : it}, umeng_event_url: Str,timestamp:Timestamp, status:{StringUtils.isBlank(it as String) ? 0 : it as Integer},
+            umeng_event: Str, umeng_event_room_id: {StringUtils.isBlank(it as String) ? null : it as Integer}, umeng_event_url: Str,timestamp:Timestamp, status:{StringUtils.isBlank(it as String) ? 0 : it as Integer},
             stime:{String str->  StringUtils.isBlank(str) ? null : Web.getTime(str).getTime()},
             etime:{String str->  StringUtils.isBlank(str) ? null : Web.getTime(str).getTime()}
     ]
@@ -185,7 +185,7 @@ class PushController extends BaseController {
                             ],
                             "extra": [
                                 "event": map.get("umeng_event"),
-                                "room_id": map.get("umeng_event_room_id") as Integer,
+                                "room_id": map.get("umeng_event_room_id"),
                                 "url": map.get("umeng_event_url")
                             ]
                     ]
